Move door code files into /runtime
