+++ lib/runtime/doors/door-command-helper.js
@@ -1,7 +1,7 @@
 // @ts-check
 'use strict';
 
-const Helper = require('../helpers/entity-resolution-helper');
+const Helper = require('../../helpers/entity-resolution-helper');
 
 const OPPOSITE_DIRECTION = Object.freeze({
   north: 'south',
