+++ lib/runtime/doors/go-exit-hook-helper.js
@@ -1,7 +1,7 @@
 // @ts-check
 'use strict';
 
-const { buildRoomViewLines } = require('../helpers/room-view-helper');
+const { buildRoomViewLines } = require('../../helpers/room-view-helper');
 
 const OPPOSITE_DIRECTION = Object.freeze({
   north: 'south',
